Update dependencies and plugins
Keep maven-archetypes to 3.1.0 (rather than the latest
3.1.2) due to 3.1.1 introducing breaking changes to
"reference" projects in integration tests, and more
importantly behaving differently depending on the
JDK version (ARCHETYPE-584, ARCHETYPE-587).
Maybe it would be enough to test with a single JDK
version only, but we're not vulnerable to
CVE-2018-1000632 (ARCHETYPE-568) anyway here, this is
more for users of archetype:generate.